need workaround for integer range, we need float increments

+++ modules/computeFX/FX.py
@@ -1,8 +1,8 @@
-def FX(Polynomial: list[list[str]], X) -> int:
+def FX(Polynomial: list[list[str]], X) -> float:
     FXSum = 0
 
     for Term in Polynomial:
-        FXSum = FXSum + (Term[0] * (X ** Term[1]))
+        FXSum = float(FXSum) + float((float(Term[0]) * (X ** float(Term[1]))))
     
     
     return FXSum
